server/pyproject.toml
[tool.poetry]
name = "playlist"
version = "0.1.0"
description = ""
authors = ["Violet Eldridge <github@violet-mica.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = "^2.0.36"
pydantic = "^2.9.2"
requests = "^2.32.3"
python-vlc = "^3.0.21203"
m3u8 = "^6.0.0"
sqlalchemy-stubs = "^0.4"


[tool.poetry.group.dev.dependencies]
ipython = "^8.28.0"
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
